// apps/web/src/generated/types/researchers/GetResearchers.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const getResearchers200GlobalRoleEnum = {
    PRINCIPAL_INVESTIGATOR: "PRINCIPAL_INVESTIGATOR",
    POSTDOC: "POSTDOC",
    GRADUATE_STUDENT: "GRADUATE_STUDENT",
    LAB_TECHNICIAN: "LAB_TECHNICIAN"
} as const;

export type GetResearchers200GlobalRoleEnumKey = (typeof getResearchers200GlobalRoleEnum)[keyof typeof getResearchers200GlobalRoleEnum];

/**
 * @description Response for status 200
*/
export type GetResearchers200 = {
    /**
     * @type string
    */
    id: string;
    /**
     * @type string
    */
    name: string;
    /**
     * @type string
    */
    email: string;
    /**
     * @type string
    */
    globalRole: GetResearchers200GlobalRoleEnumKey;
    /**
     * @type number
    */
    projectCount: number;
    /**
     * @type number
    */
    measurementCount: number;
}[];

export type GetResearchersQueryResponse = GetResearchers200;

export type GetResearchersQuery = {
    Response: GetResearchers200;
    Errors: any;
};